requirements.txt
Brotli==1.0.9
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.2
Flask-Compress==1.10.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
paho-mqtt==1.6.1
plotly==5.3.1
six==1.16.0
tenacity==8.0.1
Werkzeug==2.0.2
wincertstore==0.2
